+++ clever-task/src/main/java/org/clever/task/core/model/entity/TaskJob.java
@@ -17,17 +17,17 @@ public class TaskJob implements Serializable {
     /** 任务名称 */
     private String name;
     /** 任务类型，1：http调用，2：java调用，3：js脚本，4：shell脚本 */
-    private Byte type;
+    private Integer type;
     /** 最大重入执行数量(对于单个节点当前任务未执行完成就触发了下一次执行导致任务重入执行)，小于等于0：表示禁止重入执行 */
-    private Byte maxReentry;
+    private Integer maxReentry;
     /** 是否允许多节点并发执行，使用悲观锁实现，不建议禁止，0：禁止，1：允许 */
-    private Byte allowConcurrent;
+    private Integer allowConcurrent;
     /** 悲观锁 */
     private Long lockVersion;
     /** 执行失败时的最大重试次数 */
     private Integer maxRetryCount;
     /** 路由策略，0：不启用，1：指定节点优先，2：固定节点白名单，3：固定节点黑名单 */
-    private Byte routeStrategy;
+    private Integer routeStrategy;
     /** 路由策略，1-指定节点优先，调度器名称集合 */
     private String firstScheduler;
     /** 路由策略，2-固定节点白名单，调度器名称集合 */
@@ -35,13 +35,13 @@ public class TaskJob implements Serializable {
     /** 路由策略，3-固定节点黑名单，调度器名称集合 */
     private String blacklistScheduler;
     /** 负载均衡策略，1：抢占，2：随机，3：轮询，4：一致性HASH */
-    private Byte loadBalance;
+    private Integer loadBalance;
     /** 是否更新任务数据，0：不更新，1：更新 */
-    private Byte isUpdateData;
+    private Integer isUpdateData;
     /** 任务数据(json格式) */
     private String jobData;
     /** 是否禁用：0-启用，1-禁用 */
-    private Byte disable;
+    private Integer disable;
     /** 描述 */
     private String description;
     /** 创建时间 */
